# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dcss Swagger API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/role": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "添加角色",
                "parameters": [
                    {
                        "description": "角色名 备注 状态 []{用户}",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReqAddRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加角色",
                        "schema": {
                            "$ref": "#/definitions/resp.Result"
                        }
                    }
                }
            }
        },
        "/auth/role/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "通过id来获取角色信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "通过id来获取角色信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.RespGetRoleByID"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "通过id来修改角色信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "名称 备注 状态 []{用户ID}",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReqEditRoleByID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "通过id来修改角色信息",
                        "schema": {
                            "$ref": "#/definitions/resp.Result"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "通过id来删除角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "通过id来删除角色",
                        "schema": {
                            "$ref": "#/definitions/resp.Result"
                        }
                    }
                }
            }
        },
        "/auth/role/{id}/api_auth": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "通过角色id来绑定Api权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ApiIds[]int",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReqSetRoleApiAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "通过角色id来绑定Api权限",
                        "schema": {
                            "$ref": "#/definitions/resp.Result"
                        }
                    }
                }
            }
        },
        "/auth/role/{id}/menu_auth": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "通过角色id来绑定菜单权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "MenuIds[]int",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReqSetRoleApiAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "通过角色id来绑定菜单权限",
                        "schema": {
                            "$ref": "#/definitions/resp.Result"
                        }
                    }
                }
            }
        },
        "/auth/role/{id}/state/": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "通过id来修改角色状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "中文名 电话 邮箱 状态",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReqEditRoleStateByID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "通过id来修改角色状态",
                        "schema": {
                            "$ref": "#/definitions/resp.Result"
                        }
                    }
                }
            }
        },
        "/auth/roles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "获取所有的角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "过滤关键字",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pagenum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "返回数量",
                        "name": "pagesize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取所有的角色",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.RespGetRoleList"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/roles/{id}/api_auth": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "根据角色id来获取api权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "根据角色id来获取api权限",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.CasbinInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/roles/{id}/menu_auth": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "根据角色id来获取菜单权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "根据角色id来获取菜单权限",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CasbinInfo": {
            "type": "object",
            "properties": {
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "models.CommonSysUser": {
            "type": "object",
            "properties": {
                "chinese_name": {
                    "type": "string",
                    "example": "中文名"
                },
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string",
                    "example": "用户名"
                }
            }
        },
        "models.Meta": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "keepAlive": {
                    "type": "boolean"
                },
                "rank": {
                    "type": "integer"
                },
                "showLink": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.ReqAddRole": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "remark": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "state": {
                    "type": "boolean"
                },
                "user_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "models.ReqEditRoleByID": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "remark": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "state": {
                    "type": "boolean"
                },
                "user_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.ReqEditRoleStateByID": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "boolean"
                }
            }
        },
...
        "models.ReqSetRoleApiAuth": {
            "type": "object",
            "required": [
                "api_infos"
            ],
            "properties": {
                "api_infos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.RespGetRoleByID": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "创建时间"
                },
                "created_by": {
                    "$ref": "#/definitions/models.CommonSysUser"
                },
                "created_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "menus": {
                    "description": "角色菜单多对多关系",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SysMenu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "state": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string",
                    "example": "更新时间"
                },
                "updated_by": {
                    "$ref": "#/definitions/models.CommonSysUser"
                },
                "updated_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CommonSysUser"
                    }
                }
            }
        },
        "models.RespGetRoleList": {
            "type": "object",
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RespGetRoleByID"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.SysMenu": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SysMenu"
                    }
                },
                "component": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "创建时间"
                },
                "created_by": {
                    "$ref": "#/definitions/models.CommonSysUser"
                },
                "created_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "meta": {
                    "$ref": "#/definitions/models.Meta"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SysRole"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "更新时间"
                },
                "updated_by": {
                    "$ref": "#/definitions/models.CommonSysUser"
                },
                "updated_id": {
                    "type": "integer"
                }
            }
        },
        "models.SysRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "创建时间"
                },
                "created_by": {
                    "$ref": "#/definitions/models.CommonSysUser"
                },
                "created_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "menus": {
                    "description": "角色菜单多对多关系",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SysMenu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "state": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string",
                    "example": "更新时间"
                },
                "updated_by": {
                    "$ref": "#/definitions/models.CommonSysUser"
                },
                "updated_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SysUser"
                    }
                }
            }
        },
...
        "resp.Result": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "成功"
                },
                "status": {
                    "type": "integer",
                    "default": 200
                },
                "timestamp": {
                    "type": "string",
                    "example": "响应时间"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "jwt 认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}